image_classification/requirements.txt
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
Pillow>=10.0.0
tqdm>=4.65.0
joblib>=1.3.0
pandas>=2.0.0
requests>=2.31.0
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

